move start handler to `@remotion/serverless`
